fix: move navigation bar to component folder
